add db sub command
